+++ source/components/cardContainer/container/columnHeader.template.ts
@@ -1,0 +1,12 @@
+import { Directive, TemplateRef, Input } from '@angular/core';
+
+@Directive({ selector: '[rlColumnHeader]' })
+export class ColumnHeaderTemplate {
+	@Input() rlColumnHeaderName: string;
+
+	get name(): string {
+		return this.rlColumnHeaderName;
+	}
+
+	constructor(public template: TemplateRef<any>) {	}
+}
